(build): introduce view and elk module
this separates the view related parts from the editor into
its own module

also separates elk into an own module
