+++ editor/js/src/main/scala/com/flowtick/graphs/editor/EditorViewJs.scala
@@ -1,15 +1,16 @@
 package com.flowtick.graphs.editor
 
 import cats.effect.IO
-import com.flowtick.graphs.editor.view.{SVGRendererJs, SVGRendererOptions}
+import com.flowtick.graphs.editor.view.SVGRendererJs
+import com.flowtick.graphs.view.SVGRendererOptions
 import org.scalajs.dom.raw.{Element, Event}
 
-class EditorViewJs(containerElementId: String)(val messageBus: EditorMessageBus)
-    extends EditorView[Element, Event] {
+class EditorViewJs(containerElementId: String)(messageBus: EditorMessageBus)
+    extends EditorView[Element, Event](messageBus) {
   lazy val container =
     org.scalajs.dom.window.document.getElementById(containerElementId)
 
-  def createPage: IO[Page[Element, Event]] = for {
+  def createPage: IO[PageType] = for {
     newPage <- IO.pure(
       EditorPageJs(handleSelect, handleDrag, handleDoubleClick)(
         SVGRendererJs(SVGRendererOptions(showOrigin = true)),
