+++ editor/shared/src/main/scala/com/flowtick/graphs/editor/EditorMain.scala
@@ -1,11 +1,9 @@
 package com.flowtick.graphs.editor
 
 import cats.effect.IO
-import cats.implicits._
-import com.flowtick.graphs.Graph
-import com.flowtick.graphs.editor.feature.{ModelUpdateFeature, RoutingFeature, UndoFeature}
-import com.flowtick.graphs.style._
 import cats.effect.kernel.Ref
+import cats.implicits._
+import com.flowtick.graphs.editor.feature.{ModelUpdateFeature, RoutingFeature, UndoFeature}
 
 final case class EditorInstance(
     bus: EditorMessageBus,
